wip: removed old token editor
