+++ src/property-filter/styles.scss
@@ -39,11 +39,13 @@ $operator-field-width: 120px;
   font-weight: bold;
 }
 
-.property-editor,
-.token-editor {
+.property-editor {
   user-select: text;
   margin-block: awsui.$space-xxs;
   margin-inline: awsui.$space-xxs;
+  padding-block: awsui.$space-m;
+  padding-inline: awsui.$space-m;
+  overflow-y: auto;
 
   &-form {
     margin-block-end: awsui.$space-scaled-l;
@@ -78,14 +80,6 @@ $operator-field-width: 120px;
 }
 
 .token-editor {
-  &-actions {
-    // The below code cancels horizontal padding of the popover and horizontal margin of the token-editor.
-    padding-inline-end: calc(#{awsui.$space-m} + #{awsui.$space-xxs});
-    margin-inline: calc(-1 * #{awsui.$space-m} + -1 * #{awsui.$space-xxs});
-  }
-}
-
-.token-editor-grouped {
   display: flex;
   flex-direction: column;
   gap: awsui.$space-s;
@@ -144,7 +138,7 @@ $operator-field-width: 120px;
       grid-template-columns: 1fr;
       gap: awsui.$space-m;
 
-      > .token-editor-grouped-group {
+      > .token-editor-group {
         display: flex;
         flex-direction: column;
         gap: awsui.$space-s;
@@ -168,12 +162,6 @@ $operator-field-width: 120px;
   &-add-token {
     margin-block-start: awsui.$space-s;
   }
-}
-
-.property-editor {
-  padding-block: awsui.$space-m;
-  padding-inline: awsui.$space-m;
-  overflow-y: auto;
 }
 
 .custom-content-wrapper {
