+++ src/property-filter/token.tsx
@@ -90,13 +90,27 @@ export const TokenButton = ({
       disabled={disabled}
       editorContent={
         <TokenEditor
+          supportsGroups={false}
           filteringProperties={filteringProperties}
           filteringOptions={filteringOptions}
-          temporaryToken={temporaryToken}
-          onChangeTemporaryToken={setTemporaryToken}
+          tempGroup={[temporaryToken]}
+          onChangeTempGroup={newGroup => setTemporaryToken(newGroup[0])}
+          // This property will be needed when supportsGroups={true}
+          standaloneTokens={[]}
+          // This property will be needed when supportsGroups={true}
+          onChangeStandalone={() => {}}
           asyncProps={asyncProps}
           onLoadItems={onLoadItems}
-          i18nStrings={i18nStrings}
+          i18nStrings={{
+            ...i18nStrings,
+            // These properties will be needed when supportsGroups={true}
+            tokenEditorTokenActionsLabel: () => '',
+            tokenEditorTokenRemoveLabel: () => '',
+            tokenEditorTokenRemoveFromGroupLabel: () => '',
+            tokenEditorAddNewTokenLabel: '',
+            tokenEditorAddTokenActionsLabel: '',
+            tokenEditorAddExistingTokenLabel: () => '',
+          }}
           asyncProperties={asyncProperties}
           customGroupsText={customGroupsText}
           freeTextFiltering={freeTextFiltering}
